src/wav.c : Fix length of 'data' section.
The WAV file 'data' chunk should always be an even number of bytes in
length. To ensure this we add a padding byte as needed. On read we
check the 'data' chunk length and write a warning to the internal log
file if the 'data' chunk is not an even number of bytes in length.

Closes: https://github.com/erikd/libsndfile/issues/61