CMake : Build system nears parity with autotool build system.
Specifically, it can now build the shared library (without Ogg/Vorbis
or FLAC support) and the programs as well as building and running
the tests using 'make check'. This new CMake build system has a single
top level build description and on Unix, by default generates a non-
recursive Makefile (unlike the existing autotools build system).

I dedicate this commit to Peter Miller (pmiller) who died in the early
hours of 2014/07/27 after a long illness. Peter was the author of the
paper "Recursive Make Considered Harmful":

    http://aegis.sourceforge.net/auug97.pdf

This paper was the inspiration for the build systems for the Linux
kernel, the Glasgow Haskell Compiler and many other projects.

Peter was also the author of a Make replacement program Cook:

    http://miller.emu.id.au/pmiller/software/cook/

Peter was a great friend, a great work colleague and an extremely
skilled and knowledgeable software engineer.

Goodbye my friend. #commitforpmiller