RF64: Implement auto RF64 downgrade to WAV
The EBU recomendation is that when writing RF64 files and the
resulting file is less than 4Gig in size, it should be downgraded
to a WAV file (WAV files have a maximum size of 4Gig).

In libsndfile we don't quite follow the EBU recommendations, mainly
because the test suite needs to be able test reading/writing RF64
files without having to generate 4 gigabyte files. Therefore, by
default, libsndfile does not do an auto-downgrade. Instead, once
an RF64 file is has been opened for write, there is a command:

    sf_command (file, SFC_RF64_AUTO_DOWNGRADE, NULL, 0);

that will switch libsndfile to work as per the EBU recomendations.
This command has to be run for each and every RF64 file for which the
EBU recommended behaviour is desired.

Original patch by Paul Davis (linuxaudiosystems.com) but tweaked
slightly.