Separate out the configure flags for libflac, libogg and libvorbis
Previously you could only configure all of libflac, libogg and libvorbis
or none of them. This change separates out the detection/enabling/disabling
of them into --enable-libflac --enable-libogg and --enable-vorbis.

Apart from being useful in itself, this makes it possible to
add other formats via external libraries (libopus, libspeex ...)
without insisting that all of the external libraries be available or none.

We keep the old option --disable-external-libs because it is used in the
build scripts for OpenWRT, OpenEmbedded and others.