src/paf.c: Fix a long standing endian issue
There has been a long standing issue where something was not quite
right with the endian-ness handle of PAF files. This was difficlt to
track down because of the difficulty in getting access to a Paris Audio
system. Fortunately, Jeremy Friesner <jfriesne@gmail.com> was able to
provide some test files and a preliminary patch.