Use UTF-8 as internal path encoding
This is an internal change to unify and simplify the handling of file
paths.

On the Windows platform, the file path is always converted to UTF-8 and
converted to UTF-16 only for calls to WinAPI functions.

The behavior of the functions for opening files on non-Windows platforms
does not change.