common: Change float->int convert funcs to accept scale factor
Change the floating to integer conversion functions in common to accept
a scale factor argument, rather than simply a boolean 'normalize'
argument. This allows the functions be used for not only the float->int
conversion of writing floating point data to an integer format, but also
the float->int conversion of reading integer data from a floating point
format.

The numerical constants uses for the clipping versions of these
functions are updated to be of type float rather than double, and the
values are changed to be representable as floats. (Eg, 0x7fffffff cannot
be represented as a 32-bit float.

The CPU-specific clipping defines are removed for the to-short
functions, as the CPU specific clipping would be for 32-bit values that
then have to be truncated to 16 bit, which very probably won't work.

The only current consumer of these functions, alac.c is updated to work
with these changes.