SetupABIVersions.cmake: Fix Libtool compatibility for lt_age >0
The related code in GNU Libtool is this:

  func_arith $current + 1
  minor_current=$func_arith_result

So the minor part does not go into minor_current in practice,
despite name "minor_current" indicating that.

The only reason the math bug didn't show earlier is because
SNDFILE_ABI_VERSION_MINOR is 0 currently, and adding 0 doesn't
change the result.  The flow of data is this:

SNDFILE_ABI_VERSION_MINOR <-- SNDFILE_LT_AGE <-- lt_age <-- 0