flac: fix reading a large number of samples at once
When reading a FLAC file, the internal function flac_read_loop is used
to read a block of samples and copy them to the application-provided
output buffer.  The 'len' argument to flac_read_loop is the total
number of samples to read (i.e., duration times number of channels.)

Various fields in the FLAC_PRIVATE structure, as well as local
variables in various internal functions, are 'unsigned', which means
that trying to read into a buffer larger than UINT_MAX will cause
problems.  To avoid these problems, each function that calls
flac_read_loop breaks up the operation into buffers of at most
0x1000000 (2^24) total samples.

However, flac_read_loop will also fail (due to the condition at line
224) if the requested 'len' is not a multiple of the number of
channels.  Consequently, if an application tried to read:

 - more than 5592405 frames at once from a three-channel FLAC file
 - more than 3355443 frames at once from a five-channel FLAC file
 - more than 2796202 frames at once from a six-channel FLAC file
 - more than 2396745 frames at once from a seven-channel FLAC file

then sf_read_* or sf_readf_* would incorrectly fail and return zero.

As a simple fix, change the internal size limit from 0x1000000 to
0x690000 (2^16 * 3 * 5 * 7).  This is still quite arbitrary, but is
guaranteed to be divisible by the number of channels (which must be
between 1 and 8); it also ought to be large enough to avoid any
performance overhead, and small enough to be easy to test.