mpeg_decode: Fix bug where bytes from mpg123_read where ignored.
Most every read api I've ever worked with return no error and positive
byte counts, or zero bytes and an end or error condition. When the end
of stream is reached, the last read will contain the last bytes, and the
next call to read returns an error indicating that end has been reached.

But not mpg123, no, when reaching the last bytes, it returns the last
bytes AND sets the error condition to done.

Assumptions make and ass of you an mumptions.