aiff: Ignore offset padding bytes when rewriting header
When rewriting an AIFF header, only rewrite the actual header bytes and
ignore the padding bytes which may exist between the header and the data
inside the SSND chunk.

Firstly, re-writing these padding bytes is needless.

Second, as these padding bytes were never read into the binheader
buffer, and thus the buffer not grown to ensure it has capacity for
them, reading them as part of the header rewrite overflows that buffer.