mpeg: fix reading of MP3 files without Xing or INFO headers
mpg123_length relies on either an info frame at the beginning of the file,
or a previous call to mpg123_scan to determine length, otherwise
it will return 0. This causes subsequent calls to sf_read to read no
samples.

As a workaround, call mpg123_scan if mpg123_length is 0 and then try to
obtain length again.