ossfuzz: fix sometimes undefined default case
When vfseek() is called with a whence that is not SEEK_SET, SEEK_CUR, or
SEEK_END, it enters a default switch-case that does not set the
new_offset variable. The function then continues on and uses that unset
variable.

Linux 3.1+ stdio allows for SEEK_DATA and SEEK_HOLE whence values,
though they aren't conceptually supported by the vfseek() function.

Handle this case more sanely (and consistently with the lseek()
function) by setting errno to EINVAL when whence is an unsupported
value, and immediately returning -1 to flag the error.

Signed-off-by: Alex Stewart <alex.stewart@ni.com>