Add support for additional strings
libsndfile processes a limited subset of the strings found in audio
files which is defined by the enum constants of the `SF_STR_` series.

This causes programs that process audio files to drop certain metadata
outside of the accessible string values when saving changed files using
libsndfile.

This commit attempts to add support for the following additional
strings in order to allow keeping important metadata when editing
files using libsndfile:

  DISCNUMBER  - When an album consists of multiple discs, songs may
                contain information about which disc a song belongs
                to. This is info important to ensure correct sorting
                of songs by applications.

  ALBUMARTIST - Allows distinguishing between the artist of an album
                and of a song.

  PERFORMER   - This tag is included in the xiph.org specification for
                vorbis and seems to be widely supported.

  LABEL       - This tag can be used to specify the label under which
                an album was published. According to xiph.org,
                ORGANIZATION is also proposed for this, but seems to be
                less popular in practice.

  ISRC        - International Standard Recording Code
                This seems to be one of the more widely-supported ways
                to assign unique IDs to recordings

The following strings were also considered for addition but not included
due to lack of common ways to represent in multiple formats:
CDDB, TRACKTOTAL, DISCTOTAL, VERSION, UPC.

ref: #982