string_test: Add FLAC identity test for new strings
This test provides test data which contain more strings than previous
versions of libsndfile supported. When decoding all strings from that
data and writing a file with the same strings, the outcome is expected
to match predefined reference test data.

In order for the reference test data to be reproducible, it is possible
to set `create_reference` to 1 to re-create them. The idea behind not
automatically performing this re-generation all of the time is to
allow running the test against versions of the library which support
a different number of strings and be able to observe it failing for
old versions and passing for new versions. Also, not re-creating the
reference allows renumbering the enum values such as long as their order
is maintained whereas re-creation of the test data at each run requires
to either (1) refer to all of the enum values for the strings
explicitly, making it impossible to run the test on versions before the
introduction of the new constants or (2) have an implicit dependency on
the enum values.

ref: #982