string_test: Add DISCNUMBER test
Previously, the new string values had only been tested in conjunction
with FLAC. While it is not particularly worthwhile to reproduce the
implementation completely within the test, there are some cases for
which the implementation differs from the other string values. One such
example is the DISCNUMBER, which for e.g. MP3 is encoded as follows:

	mpeg_l3_encoder_add_id3v2_tag (psf, "TPOS", value) ;

This is noteworthy because the construct with `add_id3v2_tag` is new
and was not used anywhere before yet. Hence it makes sense to ensure
that this new function is also triggered by the test cases.

This commit thus adds the DISCNUMBER tests for all formats but AIFF and
CAF for which this string is not available.

ref: #982