Support reading and writing the description field in AU files
AU files support an optional description field between the header
and the audio data. This patch adds five SFC_ command codes for
manipulating the description field: two commands to get and set the
length of the field; one to write string data; one to write binary
data; and one to read the stored description.

This changes the default behavior of libsndfile when creating an AU
file, so that the data offset is 28 rather than 24. It remains
possible to create a file with 24-byte offset, by setting the
description length to zero.

This resolves issue#1087.